Crate serde_command_opts

A crate for serializing command line options (structopt in reverse).

Structs

Serializer

A serializer intended for converting objects into command options..

Enums

BooleanType

Mode to use for serializing boolean values.

Error

Errors emitted during the serialization process.

Functions

to_args

Serialize an object into command options, using the default serializer settings.